+++ src/Request/People.php
@@ -1577,7 +1577,8 @@ class People extends RequestCollection
     {
         return $this->ig->request('banyan/banyan/')
             ->addParam('is_private_share', false)
-            ->addParam('views', ($nullState) ? '["direct_ibc_nullstate"]' : '["reshare_share_sheet","direct_user_search_keypressed","story_share_sheet","direct_user_search_nullstate","direct_inbox_active_now","forwarding_recipient_sheet","call_recipients"]')
+            ->addParam('views', ($nullState) ? '["direct_ibc_nullstate"]' : '["reshare_share_sheet","direct_user_search_keypressed","story_share_sheet","direct_user_search_nullstate","direct_inbox_active_now","forwarding_recipient_sheet","call_recipients","direct_ibc_inbox_discovery"]')
+            ->addParam('IBCShareSheetParams', json_encode(['size' => 3]))
             ->addParam('is_real_time', false)
             ->getResponse(new Response\SharePrefillResponse());
     }
@@ -1616,6 +1617,7 @@ class People extends RequestCollection
      *
      * @param string $userId      Numerical UserPK ID.
      * @param string $surfaceType Platform.
+     * @param string $entrypoint  Entry point.
      *
      * @throws \InstagramAPI\Exception\InstagramException
      *
@@ -1623,9 +1625,11 @@ class People extends RequestCollection
      */
     public function getCreatorInfo(
         $userId,
-        $surfaceType = 'android')
+        $surfaceType = 'android',
+        $entrypoint = 'self_profile')
     {
         return $this->ig->request('creator/creator_info')
+            ->addParam('entry_point', $entrypoint)
             ->addParam('surface_type', $surfaceType)
             ->addParam('user_id', $userId)
             ->getResponse(new Response\UserInfoResponse());